fix: Tier 2+3 bugs and perf improvements
Bugs:
- B20: GBDT child index — int16_t→int32_t for multi-tree models
- B1:  Wire g_interp_entry in vtx_interp_run (deopt was non-functional)
- B21: Trace selector Phase 2 (already fixed by subagent)

Performance:
- Perf 2: Enable Phi in can_produce_raw_int — loop-carried SMI tag elision
  eliminates tag/untag on every loop iteration for counters/accumulators
- Perf 4: Unreserve RAX/RDX for non-IDIV functions — 2 extra registers
  available for register allocation, reducing spills in loop-heavy code
- Perf 5: Wire vtx_branch_optimize — merge inverted branch chains,
  eliminate dead branches, compute short-jump offsets

6 files changed. All compile clean.

+++ src/inliner/inference.h
@@ -73,8 +73,8 @@
 typedef struct {
     uint16_t feature_index;  /* feature to test, or VTX_GBDT_LEAF_MARKER for leaf */
     float    threshold;      /* decision threshold for split */
-    int16_t  left_child;     /* index of left child (feature < threshold) */
-    int16_t  right_child;    /* index of right child (feature >= threshold) */
+    int32_t  left_child;     /* index of left child (feature < threshold) */
+    int32_t  right_child;    /* index of right child (feature >= threshold) */
     float    leaf_value;     /* value if this is a leaf node */
 } vtx_gbdt_node_t;
